requirements.txt
opencv-python
ultralytics
deep_sort_realtime
numpy
pandas
torch
PyYAML
scipy
filterpy
